requirements.txt
pip==24.0
python-dotenv==1.0.1
reflex==0.4.2
supabase==2.4.0
